Moved logic from BasicField to a post repository, so the fields can be independent of the actual data fetching implementation
