+++ src/Field/BasicField.php
@@ -9,6 +9,7 @@ use Corcel\Model\Term;
 use Corcel\Model\Meta\TermMeta;
 use Corcel\Model\User;
 use Corcel\Model\Meta\UserMeta;
+use Corcel\Acf\Repositories\Repository;
 
 /**
  * Class BasicField.
@@ -18,148 +19,51 @@ use Corcel\Model\Meta\UserMeta;
 abstract class BasicField
 {
     /**
-     * @var Model
+     * @var Repository
      */
-    protected $post;
-
-    /**
-     * @var PostMeta
-     */
-    protected $postMeta;
-
-    /**
-     * @var string
-     */
-    protected $name;
-
-    /**
-     * @var string
-     */
-    protected $key;
-
-    /**
-     * @var string
-     */
-    protected $type;
-
-    /**
-     * @var mixed
-     */
-    protected $value;
-
-    /**
-     * @var string
-     */
-    protected $connection;
+    protected $repository;
 
     /**
      * Constructor method.
      *
-     * @param Post $post
+     * @param Repository $repository
      */
-    public function __construct(Model $post)
+    public function __construct(Repository $repository)
     {
-        $this->post = $post;
-
-        if ($post instanceof Post) {
-            $this->postMeta = new PostMeta();
-        } elseif ($post instanceof Term) {
-            $this->postMeta = new TermMeta();
-        } elseif ($post instanceof User) {
-            $this->postMeta = new UserMeta();
-        }
-
-        $this->postMeta->setConnection($post->getConnectionName());
+        $this->repository = $repository;
     }
 
     /**
-     * Get the value of a field according it's post ID.
-     *
-     * @param string $field
-     *
-     * @return array|string
+     * @deprecated in favor of $this->repository->fetchValue()
      */
     public function fetchValue($field)
     {
-        $postMeta = $this->postMeta->where(
-           $this->getKeyName(), $this->post->getKey()
-        )->where('meta_key', $field)->first();
-
-        if (isset($postMeta->meta_value) and ! is_null($postMeta->meta_value)) {
-            $value = $postMeta->meta_value;
-            if ($array = @unserialize($value) and is_array($array)) {
-                $this->value = $array;
-
-                return $array;
-            } else {
-                $this->value = $value;
-
-                return $value;
-            }
-        }
+        return $this->repository->fetchValue($field);
     }
 
     /**
-     * @param string $fieldName
-     *
-     * @return string
+     * @deprecated in favor of $this->repository->fetchValue()
      */
     public function fetchFieldKey($fieldName)
     {
-        $this->name = $fieldName;
-
-        $postMeta = $this->postMeta->where($this->getKeyName(), $this->post->getKey())
-            ->where('meta_key', '_' . $fieldName)
-            ->first();
-
-        if (!$postMeta) {
-            return null;
-        }
-
-        $this->key = $postMeta->meta_value;
-
-        return $this->key;
+        return $this->repository->fetchFieldKey($fieldName);
     }
 
     /**
-     * @param string $fieldKey
-     *
-     * @return string|null
+     * @deprecated in favor of $this->repository->fetchValue()
      */
     public function fetchFieldType($fieldKey)
     {
-        $post = Post::on($this->post->getConnectionName())
-                   ->orWhere(function ($query) use ($fieldKey) {
-                       $query->where('post_name', $fieldKey);
-                       $query->where('post_type', 'acf-field');
-                   })->first();
-
-        if ($post) {
-            $fieldData = unserialize($post->post_content);
-            $this->type = isset($fieldData['type']) ? $fieldData['type'] : 'text';
-
-            return $this->type;
-        }
-
-        return null;
+        return $this->repository->fetchFieldType($fieldKey);
     }
 
     /**
-     * Get the name of the key for the field.
-     *
-     * @return string
+     * @deprecated in favor of $this->repository->fetchValue()
      */
     public function getKeyName()
     {
-        if ($this->post instanceof Post) {
-            return 'post_id';
-        } elseif ($this->post instanceof Term) {
-            return 'term_id';
-        } elseif ($this->post instanceof User) {
-            return 'user_id';
-        }
+        return $this->repository->getKeyName();
     }
-
 
     /**
      * @return mixed
