+++ src/Repositories/Repository.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Corcel\Acf\Repositories;
+
+class Repository
+{
+    public function __construct()
+    {
+    }
+}
